+++ src/Facades/FormElement.php
@@ -36,6 +36,7 @@ use Illuminate\Support\Facades\Facade;
  * @method static \SleepingOwl\Admin\Form\Element\Upload upload($name, $label = null)
  * @method static \SleepingOwl\Admin\Form\Element\Number number($name, $label = null)
  * @method static \SleepingOwl\Admin\Form\Related\Forms\HasMany hasMany(string $relationName, array $elements)
+ * @method static \SleepingOwl\Admin\Form\Related\Forms\HasManyLocal hasManyLocal(string $fieldName, array $elements)
  * @method static \SleepingOwl\Admin\Form\Related\Forms\ManyToMany manyToMany(string $relationName, array $elements)
  * @method static \SleepingOwl\Admin\Form\Related\Forms\BelongsTo belongsTo(string $relationName, array $elements)
  */